refactor migration dir structure, fixed up constructors based on review
